[FE] refactor: Album 상세 페이지의 track card 변경
