+++ frontend/components/DetailPage/index.tsx
@@ -1,13 +1,13 @@
 import React, { FC } from "react";
 import styled from "styled-components";
 import DetailHeader from "./DetailHeader";
-import { Track } from "../../interfaces";
-import TrackList from "../Tracklist";
+import { TrackProps } from "../../interfaces";
+import TrackCard from "../TrackCard";
 
 interface Props {
   type: "album" | "playlist" | "artist" | "magazine" | "news";
   detailData: any;
-  tracks: Track[];
+  tracks: TrackProps[];
 }
 
 const StyleDetailPage = styled.div`
@@ -19,7 +19,7 @@ const DetailPage: FC<Props> = ({ type, detailData, tracks }: Props) => {
   return (
     <StyleDetailPage>
       <DetailHeader type={type} detailData={detailData} tracks={tracks} />
-      <TrackList Tracks={tracks} />
+      <TrackCard data={tracks} />
     </StyleDetailPage>
   );
 };
