+++ frontend/interfaces/index.ts
@@ -12,6 +12,36 @@ interface Album {
   artistId: number;
 }
 
+interface TrackCardProps {
+  trackName: string;
+  cover: string;
+  artists: string[];
+  albumName: string;
+}
+
+interface TrackArtistProps {
+  id: number;
+  trackId: number;
+  artistId: number;
+  Artists: {
+    artistName: string;
+  };
+}
+
+interface TrackAlbumProps {
+  cover: string;
+  albumName: string;
+}
+
+interface TrackProps {
+  id: number;
+  trackName: string;
+  albumTrackNumber: number;
+  albumId: number;
+  Albums: TrackAlbumProps;
+  Artists_Tracks: TrackArtistProps[];
+}
+
 interface Track {
   id: number;
   trackName: string;
@@ -46,4 +76,15 @@ interface News {
   playlistId?: number;
 }
 
-export type { Track, Artist, Album, Playlist, Magazine, News };
+export type {
+  Track,
+  Artist,
+  Album,
+  Playlist,
+  Magazine,
+  News,
+  TrackCardProps,
+  TrackArtistProps,
+  TrackAlbumProps,
+  TrackProps,
+};
